skool edits 6 (add exponents, optimize keyboard & add more functions, optimize other stuff to)

+++ src/vulkan/window.hpp
@@ -13,7 +13,7 @@ public:
     ~AppWindow();
 
     bool shouldClose() { return glfwWindowShouldClose(window); }
-    VkExtent2D getExtent() { return { static_cast<uint32_t>(width), static_cast<uint32_t>(height) }; }
+    VkExtent2D getExtent() { return { width, height }; }
     bool wasWindowResized() { return framebufferResized; }
     void resetWindowResizedFlag() { framebufferResized = false; }
     void createWindowSurface(VkInstance instance, VkSurfaceKHR* surface);
@@ -22,10 +22,10 @@ public:
 
 private:
     void initWindow();
-    static void framebufferResizeCallback(GLFWwindow* window, int width, int height);
+    static void framebufferResizeCallback(GLFWwindow* window, unsigned int width, unsigned int height);
 
-    int width;
-    int height;
+    unsigned int width;
+    unsigned int height;
     bool framebufferResized = false;
     GLFWwindow* window;
 };
